Revert "Consolidated attr_volatile from c-r and thread_safe."
This reverts commit dbb75a1195cfe191a1bf11ff1bc794b95a5ab6fb.
